fix(armor): check caused incorrect padding
You can have a match on all chrs in the end marker map causing the
buffer to be sliced when it shouldn't, the easier approach is to just
make sure there's a newline before the marker. The only time this can
happen is when you're on the last line and you've matched all the marker
chrs.